// frontend/src/app/api/model/passwordPolicyRule.ts
// File generated using spec-to-code.sh, DO NOT EDIT MANUALLY.

/**
 * ADH6 API
 *
 * Contact: equipe@minet.net
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PasswordPolicyRule { 
    /**
     * Keycloak password-policy provider identifier
     */
    name: string;
    /**
     * Configured provider value, or an empty string for a value-less rule
     */
    value: string;
}